Added Reactive<AudioProcessor> and tests
