+++ Tests/Source/Tests/ReactiveGUITest.cpp
@@ -1,66 +1,5 @@
 #include "../Other/TestPrefix.h"
 
-
-TEST_CASE("Reactive<Value> conversion",
-          "[Reactive<Value>][ValueExtension]")
-{
-    Reactive<Value> value;
-
-    IT("supports copy assignment from var-compatible types")
-    {
-        value = 3;
-        value = Array<var>({ 6, 7, 5 });
-        value = "Some String";
-        REQUIRE(value.getValue() == "Some String");
-    }
-
-    IT("can be converted to var")
-    {
-        value.setValue("Testing");
-        var v = value.operator var();
-        REQUIRE(v == "Testing");
-    }
-
-    IT("supports == and != with var-compatible types")
-    {
-        value.setValue("Hello!");
-        REQUIRE(value == "Hello!");
-        REQUIRE(value != "World");
-        REQUIRE(value != 3.45);
-        REQUIRE(value != 2);
-    }
-}
-
-
-TEST_CASE("Reactive<Value> Observable",
-          "[Reactive<Value>][ValueExtension]")
-{
-    auto value = std::make_shared<Reactive<Value>>("Initial");
-    Array<var> items;
-    varxCollectItems(value->rx.subject, items);
-
-    IT("emits items asynchronously when the Value changes")
-    {
-        value->setValue("Second");
-        varxCheckItems(items, "Initial");
-        varxRunDispatchLoop();
-        varxCheckItems(items, "Initial", "Second");
-        value->setValue("Third");
-        varxRunDispatchLoop();
-
-        varxRequireItems(items, "Initial", "Second", "Third");
-    }
-
-    IT("stops emitting items immediately when being destroyed")
-    {
-        value->setValue("Should not arrive");
-        value.reset();
-        varxRunDispatchLoop();
-
-        varxRequireItems(items, "Initial");
-    }
-}
-
 TEST_CASE("Reactive<Component>",
           "[Reactive<Component>][ComponentExtension]")
 {
