+++ varx/integration/varx_ExtensionBase.cpp
@@ -1,0 +1,9 @@
+ExtensionBase::ExtensionBase()
+: _deallocated(1),
+deallocated(_deallocated) {}
+
+ExtensionBase::~ExtensionBase()
+{
+    _deallocated.onNext(var::undefined());
+    _deallocated.onCompleted();
+}
