+++ varx/integration/varx_ReactiveGUI.h
@@ -1,64 +1,24 @@
 #pragma once
 
 namespace detail {
-    template<typename Base, typename T>
-    using Is = typename std::enable_if<std::is_base_of<Base, T>::value>::type;
+template<typename Base, typename T>
+using Is = typename std::enable_if<std::is_base_of<Base, T>::value>::type;
 
-    template<typename T>
-    using IsImageComponent = typename std::enable_if<std::is_base_of<juce::ImageComponent, T>::value>::type;
+template<typename T>
+using IsImageComponent = typename std::enable_if<std::is_base_of<juce::ImageComponent, T>::value>::type;
 
-    template<typename T>
-    using IsSimpleComponent = typename std::enable_if<std::is_base_of<juce::Component, T>::value && !std::is_base_of<juce::ImageComponent, T>::value && !std::is_base_of<juce::Button, T>::value && !std::is_base_of<juce::Label, T>::value && !std::is_base_of<juce::Slider, T>::value>::type;
+template<typename T>
+using IsSimpleComponent = typename std::enable_if<std::is_base_of<juce::Component, T>::value && !std::is_base_of<juce::ImageComponent, T>::value && !std::is_base_of<juce::Button, T>::value && !std::is_base_of<juce::Label, T>::value && !std::is_base_of<juce::Slider, T>::value>::type;
 
-    template<typename T>
-    using IsButton = typename std::enable_if<std::is_base_of<juce::Button, T>::value>::type;
+template<typename T>
+using IsButton = typename std::enable_if<std::is_base_of<juce::Button, T>::value>::type;
 
-    template<typename T>
-    using IsLabel = typename std::enable_if<std::is_base_of<juce::Label, T>::value>::type;
+template<typename T>
+using IsLabel = typename std::enable_if<std::is_base_of<juce::Label, T>::value>::type;
 
-    template<typename T>
-    using IsSlider = typename std::enable_if<std::is_base_of<juce::Slider, T>::value>::type;
+template<typename T>
+using IsSlider = typename std::enable_if<std::is_base_of<juce::Slider, T>::value>::type;
 }
-
-// If you get an error here, it means that you are trying to use an unsupported type T.
-template<typename T, typename Enable = void>
-class Reactive;
-
-/**
-    Adds reactive extensions to a juce::Value.
- 
-    Instead of creating a juce::Value, create an instance of this as follows:
- 
-        Reactive<Value> myValue;
- 
-    It inherits from juce::Value, so you can use it as a drop-in replacement. And you can access `myValue.rx.subject` to subscribe to changes, etc.:
- 
-        myValue.rx.subject.map(...).filter(...).subscribe(...);
- 
-    And you can subscribe this value to some Observable, to change the value whenever the Observable emits an item:
-    
-        someObservable.subscribe(myValue.rx.subject);
- */
-template<>
-class Reactive<juce::Value> : public juce::Value
-{
-public:
-    /** Creates a new instance. Has the same behavior as the juce::Value equivalent. */
-    ///@{
-    Reactive();
-    Reactive(const juce::Value& other);
-    explicit Reactive(const juce::var& initialValue);
-    ///@}
-
-    /** Sets a new value. This is the same as calling Observed<Value>::setValue. */
-    Reactive& operator=(const juce::var& newValue);
-
-    /** The reactive extension object. */
-    const ValueExtension rx;
-
-private:
-    Reactive& operator=(const Reactive&) = delete;
-};
 
 /**
     Adds reactive extensions to a juce::Component (or subclass).
